fix: better division of db and sheet frontends
allows db development without touching any of the sheet frontend
